refactor: Move authentication logic from user to auth domain (#193)
- Moved authentication-related responsibilities, including sign-up and phone verification, from the `user` domain to the `auth` domain.
- Migrated relevant data transfer objects (DTOs) from the `data.user.dto` package to `data.auth.dto`.
- Relocated API endpoints for `signUp`, `sendPhoneCode`, and `verifyPhoneCode` from `UserApiService` to `AuthApiService`.
- Shifted the corresponding methods from `UserRemoteDataSource` and `UserRepository` to `AuthRemoteDataSource` and `AuthRepository`.
- Updated `JoinNicknameViewModel` and `JoinPhoneVerificationViewModel` to inject and use `AuthRepository` instead of `UserRepository` for handling authentication flows.
